requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
matplotlib>=3.7.0
plotly>=5.14.0
scipy>=1.10.0
numba>=0.57.0
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
